pff_cardinals_scraper_repo/requirements.txt
pandas
beautifulsoup4
gspread
google-auth
playwright